% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{Internal-objects}
\alias{Internal-objects}
\alias{.add_internal_blocks}
\title{This man page is automatically generated to gather all internal functions, classes and objects that are
not meant to be called outside of the package.}
\usage{
.add_internal_blocks(blocks, env, base_path = env)
}
\description{
This man page is automatically generated to gather all internal functions, classes and objects that are
not meant to be called outside of the package.
}
\keyword{internal}
