% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-namespace2.R
\name{roclet_process.roclet_namespace2}
\alias{roclet_process.roclet_namespace2}
\title{Alternative NAMESPACE Roclet}
\usage{
\method{roclet_process}{roclet_namespace2}(x, blocks, env, base_path = env)
}
\arguments{
\item{x}{A \code{roclet} object.}

\item{blocks}{A list of \link[roxygen2]{roxy_block} objects.}

\item{env}{Package environment.}

\item{base_path}{Path to root of source package.}
}
\description{
Same as \link[roxygen2:roclet_process.roclet_namespace]{roxygen2::roclet_process.roclet_namespace}, but allow for keeping the
NAMESPACE entries unsorted.
This is useful to avoid conflicts in generics.
}
