% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-rd2.R
\name{rd2_roclet}
\alias{rd2_roclet}
\alias{roclet_tags.roclet_rd2}
\alias{roclet_process.roclet_rd2}
\alias{roclet_clean.roclet_rd2}
\title{Replacement Roclet for Generating Rd Files}
\usage{
rd2_roclet()

\method{roclet_tags}{roclet_rd2}(x)

\method{roclet_process}{roclet_rd2}(x, parsed, base_path,
  global_options = list())

\method{roclet_clean}{roclet_rd2}(x, base_path)
}
\description{
This roclets acts as a drop-in replacement for [roxygen2::rd_roclet], 
adding some extra features such as bibliography, or inline declaration
for S4 generic.
}
\details{
To use it add it needs to be declared in the roxygen options within 
your package's DESCRIPTION file:
```
Roxygen: list(roclets = c('collate', 'namespace', 'roclets::rd2_roclet'))
```


Extra features are sometimes submitted as pull request to be 
incorporated into the main `roxygen2` package.
This roclet enables using while they are being reviewed -- and 
potentially not accepted.
}
\section{Roxygen tags and commands}{

  * @cite: 
  * @bibliography:
  * \\cite: 
  * @inline:
}

