% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{rd2_tag}
\alias{rd2_tag}
\alias{roxy_tag_parse.roxy_tag_cite}
\alias{roxy_tag_parse.roxy_tag_bibliography}
\title{Extra Roxygen Tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_cite}(x)

\method{roxy_tag_parse}{roxy_tag_bibliography}(x)
}
\arguments{
\item{x}{roxy_tag object}
}
\description{
Extra Roxygen Tags
}
\section{Functions}{
\itemize{
\item \code{roxy_tag_parse.roxy_tag_cite}: Provides support for tags \verb{@cite} that should contain Bibtex
entry keys, which are looked-up in the \code{inst/REFERENCES.bib} file or external
Bibtex files specified via tag @bibliography

\item \code{roxy_tag_parse.roxy_tag_bibliography}: Provides support for tag \verb{@bibliography}
}}

